The CRT beam now has a gaussian shape
Modified the pixel shader so that the formula simulates a gaussian-shaped beam instead of a disc-shaped beam. A gaussian shaped CRT beam is physically more plausible and looks better. The downside are possible performance implications: The shader is a bit more complicated, and because the beam intensity is concentrated in the middle, we need to draw the lines twice as wide to achieve the same effect.