Render points in addition to lines
The renderer is now able to render points. (This is necessary because lines of length 0 would be invisible -- the intuition is that the cathode ray sweep across a line at constant speed, but rests at a point for a defined amount of time.) Use this to render ':' characters in the menu.