Simplified match initialization and update
The player class no longer has Initialize() and Update() methods -- this was too confusing and ultimately not useful enough. Instead, only the Match is now initialized and updated; if you need to do anything specific per player, you have to write the loop yourself.

As a convenience, there's now a Match.Graphics object. Everything drawn onto that will be shown to every player.