Game input refactoring
There's now a clear distinction between the current state of a key (up or down) and button activation events. A key is "activated" the frame it is pressed down, and if it continues to be pressed, repeated every X frames.

This new activation method is used to make moving pieces in "Falling Blocks" and laying bombs in "Bomberguy" more robust and intuitive.