Refactoring: Moved Graphics into Output
The Graphics object is now part of the new Output object, accessed by (for example) player.Output.Graphics. This is a breaking change for existing games, which have been updated accordingly.

The new principle is that Output is an object that captures (and stores) the games output, including line graphics, particles and later also audio. For non-local players, the output is set to the "NullOutput", and everything drawn there is just discarded. In the render loop, the Output of the most recent game frame is drawn to the screen.

This new division of responsibilities simplifies the code a bit. For example, the Output object is now constant throughout a match, making it unnecessary to root a graphics object everywhere all the time.