Simplified highlighting the local player
The Player base class now has a new property, LocalColor, which is white for the local player and equal to Color for all other players. This simplifies the code for the typical case of a game with a shared screen; it's no longer necessary to draw each player's screen separately.